enable/disable the map with tab
* Created frame.c
* Include stdbool for bool variable and true/false
* By default the map is print
