+++ Makefile
@@ -51,9 +51,10 @@ SRCD	=	srcs
 SRCS	=	$(SRCD)/main.c
 SRCS	+=	$(SRCD)/check_map.c
 SRCS	+=	$(SRCD)/close.c
-SRCS	+=	$(SRCD)/parsing.c
+SRCS	+=	$(SRCD)/frame.c
 SRCS	+=	$(SRCD)/init.c
 SRCS	+=	$(SRCD)/mini_map.c
+SRCS	+=	$(SRCD)/parsing.c
 SRCS	+=	$(SRCD)/player.c
 SRCS	+=	$(SRCD)/ray.c
 SRCS	+=	$(SRCD)/textures.c
